Add files via upload
License plate detection and recognition is the technology that uses computer vision to detect and recognize a license plate from an input image of a car. On roads, it is used to identify the cars that are breaking the traffic rules. In security, it is used to capture the license plates of the vehicles getting into and out of certain premises. In parking lots, it is used to capture the license plates of the cars being parked. The list of its applications goes on and on.